Move ITypeSymbol
